+++ src/player.rs
@@ -26,8 +26,8 @@ fn try_move_player(delta_x: i32, delta_y: i32, ecs: &mut World, world_size: Posi
     for (_player, pos) in (&mut players, &mut positions).join() {
         let destination_idx = xy_idx((pos.x + delta_x) as usize, (pos.y + delta_y) as usize);
         if map[destination_idx] != TileType::Wall {
-            pos.x = min((world_size.x - 1) as i32, max(0, pos.x + delta_x));
-            pos.y = min((world_size.y - 1) as i32, max(0, pos.y + delta_y));
+            pos.x = min((world_size.x - 1), max(0, pos.x + delta_x));
+            pos.y = min((world_size.y - 1), max(0, pos.y + delta_y));
         }
     }
 }
